fix: Properly match simulator udid if webDriverAgentUrl is provided
