+++ lib/driver.js
@@ -33,6 +33,7 @@ import {
   installToRealDevice,
   runRealDeviceReset,
   applySafariStartupArgs,
+  detectUdid,
 } from './real-device-management';
 import {
   RealDevice,
@@ -53,7 +54,6 @@ import {
   UDID_AUTO,
   checkAppPresent,
   clearSystemFiles,
-  detectUdid,
   getAndCheckIosSdkVersion,
   getAndCheckXcodeVersion,
   getDriverInfo,
@@ -1165,7 +1165,7 @@ export class XCUITestDriver extends BaseDriver {
     this.lifecycleData.createSim = false;
 
     // if we get generic names, translate them
-    this.opts.deviceName = translateDeviceName(this.opts.platformVersion, this.opts.deviceName);
+    this.opts.deviceName = translateDeviceName(this.opts.platformVersion ?? '', this.opts.deviceName);
 
     const setupVersionCaps = async () => {
       this._iosSdkVersion = await getAndCheckIosSdkVersion();
@@ -1182,7 +1182,7 @@ export class XCUITestDriver extends BaseDriver {
     if (this.opts.udid) {
       if (this.opts.udid.toLowerCase() === UDID_AUTO) {
         try {
-          this.opts.udid = await detectUdid();
+          this.opts.udid = await detectUdid.bind(this)();
         } catch (err) {
           // Trying to find matching UDID for Simulator
           this.log.warn(
@@ -1203,24 +1203,25 @@ export class XCUITestDriver extends BaseDriver {
       } else {
         // If the session specified this.opts.webDriverAgentUrl with a real device,
         // we can assume the user prepared the device properly already.
-        if (this.opts.webDriverAgentUrl) {
-          this.log.debug('Skipping checking of the device availability since the session specifies appium:webDriverAgentUrl');
+        let isRealDeviceUdid = false;
+        if (!this.opts.webDriverAgentUrl) {
+          this.log.debug(
+            'Skipping checking of the real devices availability since the session specifies appium:webDriverAgentUrl'
+          );
         } else {
-          // make sure it is a connected device. If not, the udid passed in is invalid
           const devices = await getConnectedDevices();
-          this.log.debug(`Available devices: ${devices.join(', ')}`);
-          if (!devices.includes(this.opts.udid)) {
-            // check for a particular simulator
-            this.log.debug(`No real device with udid '${this.opts.udid}'. Looking for a simulator`);
-            try {
-              const device = await getSimulator(this.opts.udid, {
-                devicesSetPath: this.opts.simulatorDevicesSetPath,
-                logger: this.log,
-              });
-              return {device, realDevice: false, udid: this.opts.udid};
-            } catch (ign) {
-              throw new Error(`Unknown device or simulator UDID: '${this.opts.udid}'`);
-            }
+          this.log.debug(`Available real devices: ${devices.join(', ')}`);
+          isRealDeviceUdid = devices.includes(this.opts.udid);
+        }
+        if (!isRealDeviceUdid) {
+          try {
+            const device = await getSimulator(this.opts.udid, {
+              devicesSetPath: this.opts.simulatorDevicesSetPath,
+              logger: this.log,
+            });
+            return {device, realDevice: false, udid: this.opts.udid};
+          } catch (ign) {
+            throw new Error(`Unknown device or simulator UDID: '${this.opts.udid}'`);
           }
         }
       }
