+++ lib/real-device-management.js
@@ -1,5 +1,6 @@
 import _ from 'lodash';
 import {buildSafariPreferences} from './app-utils';
+import {utilities} from 'appium-ios-device';
 
 /**
  * @typedef {Object} InstallOptions
@@ -100,5 +101,25 @@ export function applySafariStartupArgs() {
 }
 
 /**
+ * @this {XCUITestDriver}
+ * @returns {Promise<string>}
+ */
+export async function detectUdid() {
+  this.log.debug('Auto-detecting real device udid...');
+  const udids = await utilities.getConnectedDevices();
+  if (_.isEmpty(udids)) {
+    throw new Error('No real devices are connected to the host');
+  }
+  const udid = _.last(udids);
+  if (udids.length > 1) {
+    this.log.info(`Multiple devices found: ${udids.join(', ')}`);
+    this.log.info(`Choosing '${udid}'. Consider settings the 'udid' capability if another device must be selected`);
+  }
+  this.log.debug(`Detected real device udid: '${udid}'`);
+  return udid;
+}
+
+/**
  * @typedef {import('./real-device').RealDevice} RealDevice}
+ * @typedef {import('./driver').XCUITestDriver} XCUITestDriver
  */
