+++ lib/utils.js
@@ -1,4 +1,3 @@
-import {utilities} from 'appium-ios-device';
 import xcode from 'appium-xcode';
 import {errors} from 'appium/driver';
 import {fs, net, util} from 'appium/support';
@@ -21,21 +20,6 @@ const XCTEST_LOG_FILES_PATTERNS = [
   /^StandardOutputAndStandardError\.txt$/i,
 ];
 const XCTEST_LOGS_CACHE_FOLDER_PREFIX = 'com.apple.dt.XCTest';
-
-async function detectUdid() {
-  log.debug('Auto-detecting real device udid...');
-  const udids = await utilities.getConnectedDevices();
-  if (_.isEmpty(udids)) {
-    throw new Error('No device is connected to the host');
-  }
-  const udid = _.last(udids);
-  if (udids.length > 1) {
-    log.warn(`Multiple devices found: ${udids.join(', ')}`);
-    log.warn(`Choosing '${udid}'. If this is wrong, manually set with 'udid' desired capability`);
-  }
-  log.debug(`Detected real device udid: '${udid}'`);
-  return udid;
-}
 
 /**
  * @privateRemarks Is the minimum version really Xcode 7.3?
@@ -98,6 +82,12 @@ function getGenericSimulatorForIosVersion(platformVersion, deviceName) {
   return result;
 }
 
+/**
+ *
+ * @param {string} platformVersion
+ * @param {string} [deviceName]
+ * @returns {string|undefined}
+ */
 function translateDeviceName(platformVersion, deviceName) {
   if (!deviceName) {
     return deviceName;
@@ -114,6 +104,10 @@ function translateDeviceName(platformVersion, deviceName) {
   return deviceNameTranslated;
 }
 
+/**
+ * @param {string[]} locations
+ * @returns {Promise<void>}
+ */
 async function clearLogs(locations) {
   log.debug('Clearing log files');
   const cleanupPromises = [];
@@ -503,7 +497,6 @@ export function normalizePlatformName(platformName) {
 }
 
 export {
-  detectUdid,
   getAndCheckXcodeVersion,
   getAndCheckIosSdkVersion,
   checkAppPresent,
